pyproject.toml
[tool.poetry]
name = "tts_provider"
version = "0.1.0"
description = ""
authors = ["yomapi <yomapi.new@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
black = "^22.10.0"
djangorestframework = "^3.14.0"
PyYAML = "^6.0"
mysqlclient = "^2.1.1"
pytest-django = "^4.5.2"
pytest-benchmark = "^4.0.0"
gTTS = "^2.2.4"
django-apscheduler = "^0.6.2"
SQLAlchemy = "^1.4.43"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
